fix variable type

+++ libheif/box.cc
@@ -2304,7 +2304,7 @@ Error Box_imir::parse(BitstreamRange& range)
 {
   //parse_full_box_header(range);
 
-  uint16_t axis = range.read8();
+  uint8_t axis = range.read8();
   if (axis & 1) {
     m_axis = MirrorDirection::Horizontal;
   }
